Python/Python - Streamlit/Streamlit Project/requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
cerebras-cloud-sdk
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10